requirements.txt
python-dotenv
langchain[llms]
openai
sentence-transformers
faiss-gpu
yt-dlp
pydub
librosa